# Extend OpenVINO INT8 ignored scope