@glenn-jocher 2024-02-07
@adamp87 awesome, thanks for the PR! All benchmarks are passing now! 

One question I had was whether we've added almost all of the ops to the ignore list. Since we have all 4 major mathematical operations, is there really anything still undergoing quantization when exporting INT8 with this PR?

Can you tell what proportion of the weights/ops are quantized? Thanks!