@adamp87 2024-02-07
To answer your question in short we only leave out from quantization the last few operations, where we calculate the pixel coordinates. In more detail:

Convolution and weights
NNCF with ignored_scope `".*/Conv.*"`, counts 127 nodes of `Conv `and `Mul`. This should be 60 Conv with SiLU activation and 7 Conv without activation. During export, we add 1 Conv to ignored_scope from the DFL, which anyways have non-trainable weights. So this is **~1.5% of the Conv and 0% of the trainable weights**.

Math Operations:
NNCF counts 72 (Add,Sub,Mul,Div) in total, where 60 should be Mul from the SiLU, 6 Add operations are in the backbone. The remaining in `Detect (model.22)` are a bit tricky, because I count in Netron 10, NNCF count 9 and it does not add up to 72. So lets assume during export we ignore 9, that is 9/72, which is **~12.5% of the Math operations**.

Considering speed with `benchmark_app -m yolov8n.xml -d CPU` from OpenVINO on my old i7-9750H CPU:
Before: 
[ INFO ] Latency:
[ INFO ]    Median:        58.26 ms
[ INFO ]    Average:       **60.52 ms**
[ INFO ]    Min:           49.73 ms
[ INFO ]    Max:           123.66 ms
After: 
[ INFO ] Latency:
[ INFO ]    Median:        59.18 ms
[ INFO ]    Average:       **60.70 ms**
[ INFO ]    Min:           50.36 ms
[ INFO ]    Max:           151.63 ms