@adrianboguszewski 2024-02-07
@adamp87 do I understand this correctly that the time before and after the quantization is the same?