@adamp87 2024-02-07
@adrianboguszewski Before and After modifying the code. So before is the main branch, after is the proposed changes. I edit my post accordingly. Sorry for the confusion