@adrianboguszewski 2024-02-07
As you are ignoring more nodes could you also show the difference in mAP and inference time before and after the quantization? I mean it for this branch (changes from this PR included)