@adrianboguszewski 2024-02-07
@adamp87 I mean inference time for FP32 and INT8 (quantized with this PR) :)