@adamp87 2024-02-07
Inference time for FP32

[ INFO ] Latency:
[ INFO ]    Median:        96.03 ms
[ INFO ]    Average:       97.60 ms
[ INFO ]    Min:           64.49 ms
[ INFO ]    Max:           173.58 ms