@glenn-jocher 2024-02-07
@adamp87 thanks for the numbers, this result looks excellent to me! I think we're safe to merge to your other PR and then into main if @adrianboguszewski agrees.

The mAP increase is significant and the speed change negligible vs main.