@glenn-jocher 2024-02-07
@adamp87 ah this is interesting, our Ultralytics Actions failed to run on this PR because it's not targeted to main. They may need an update to handle this use case.